+++ CHANGELOG.md
@@ -8,6 +8,12 @@
 
 ## Unreleased (4.x)
 
+## Added
+- Added Transfers service, and moves existing after complete order handling into it
+- Added a console command for manually re-trying Stripe transfers on a Craft order
+
+## 4.0.0 - 2025-05-22
+
 > [!NOTE]
 > [Please read the upgrade guide before upgrading Marketplace.](https://craft-marketplace.kennethormandy.com/docs/getting-started/upgrading-from-v1/) The changelog documents all public API changes, but for most use cases only a few changes from the upgrade guide are required.